Move ScriptDecodeError to Cardano.CLI.Types.Errors.ScriptDecodeError
