+++ cardano-cli/src/Cardano/CLI/EraBased/Run/Certificate.hs
@@ -22,9 +22,9 @@ import           Cardano.Api
 import qualified Cardano.Api.Ledger as Ledger
 import           Cardano.Api.Shelley
 
-import           Cardano.CLI.EraBased.Errors.StakeAddress
 import           Cardano.CLI.Read
 import           Cardano.CLI.Types.Common
+import           Cardano.CLI.Types.Errors.StakeAddress
 import           Cardano.CLI.Types.Key
 
 import           Control.Monad.Trans.Class
