+++ cardano-cli/src/Cardano/CLI/Types/Errors/ScriptDecodeError.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE LambdaCase #-}
 
-module Cardano.CLI.EraBased.Errors.Common
+module Cardano.CLI.Types.Errors.ScriptDecodeError
   ( ScriptDecodeError(..)
   ) where
